Use cljc files to prepare for clojurescript support
Renames the core and core test files to cljc so reader conditionals can
be added. The rename operation is separated into a separate commit to
cleanly show the history of the test file. The added reader conditionals
in the test file would otherwise mark the file as deleted and a new one
created.

Signed-off-by: Snorre Magnus Davøen <snorre.davoen@nsd.no>